+++ cluster.c
@@ -1,4 +1,5 @@
 #include <string.h>
+#include "./generic.h"
 #include "./cluster.h"
 #include "./net.h"
 #include "./command.h"
