+++ command_raw.h
@@ -1,0 +1,6 @@
+#ifndef COMMAND_RAW_H_
+#define COMMAND_RAW_H_
+
+int commandWithRawData(Conn *, const void *, Reply *, int);
+
+#endif // COMMAND_RAW_H_
