+++ net.h
@@ -1,8 +1,6 @@
 #ifndef NET_H_
 #define NET_H_
 
-#include "./generic.h"
-
 int createConnection(Conn *);
 int createConnectionFromStr(const char *, Conn *);
 int reconnect(Conn *);
